refactor(walk_safe): move to standard util module
